ESS_ERP_DATA/ERP_OCR_AI_MODULE/requirements.txt
pytesseract
pillow
pandas>=2.0.0
fastapi
uvicorn
python-multipart
openpyxl
numpy
scikit-learn
statsmodels
opencv-python
pydantic
pdf2image